feat(pre-s6-cleanup): scaffold BigInteger migration, is_active check, pagination (TD-016/011/003)
- entry.py: Integer → BigInteger for amount column (TD-016)
- alembic/env.py: add compare_type=True so type changes are autogenerated
- migration: alter entries.amount INTEGER → BIGINT (TD-016)
- transaction_service.py: placeholder for is_active filter (TD-011)
- routes/transactions.py: placeholder for limit/offset params (TD-003)
- tests: stubs for inactive-account 422 and pagination behaviour

Note: TD-009 (root main.py) already absent from repo — will resolve in tech-debt.md.

+++ app/models/entry.py
@@ -52,7 +52,7 @@ class Entry(Base):
         Enum(Direction, name="direction"),
         nullable=False,
     )
-    amount: Mapped[int] = mapped_column(Integer, nullable=False)
+    amount: Mapped[int] = mapped_column(BigInteger, nullable=False)
     currency: Mapped[str] = mapped_column(
         String(3),
         nullable=False,  # ISO 4217 code e.g. "EUR", "USD", "JPY"
